Renamed to no-unnecessary-type-arguments
