+++ packages/eslint-plugin/docs/rules/no-unnecessary-type-arguments.md
@@ -1,4 +1,4 @@
-# Enforces that types will not to be used (use-default-type-parameter)
+# Enforces that types will not to be used (no-unnecessary-type-arguments)
 
 Warns if an explicitly specified type argument is the default for that type parameter.
